move internal dir to lupa-otelcol root level and moduling it
