+++ lupa-otelcol/exporter/opensearchexporter/go.mod
@@ -3,9 +3,9 @@ module github.com/epsagon/lupa/lupa-otelcol/exporter/opensearchexporter
 go 1.19
 
 require (
+	github.com/epsagon/lupa/lupa-otelcol/internal/modeltranslator v0.0.0-00010101000000-000000000000
 	github.com/epsagon/lupa/model v0.0.0-20221116145245-cd3200414333
 	github.com/opensearch-project/opensearch-go v1.1.0
-	github.com/stretchr/testify v1.8.1
 	go.opentelemetry.io/collector v0.64.1
 	go.opentelemetry.io/collector/pdata v0.66.0
 	go.uber.org/multierr v1.8.0
@@ -14,7 +14,6 @@ require (
 
 require (
 	github.com/cenkalti/backoff/v4 v4.1.3 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
@@ -24,7 +23,6 @@ require (
 	github.com/mitchellh/reflectwalk v1.0.2 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
 	go.opencensus.io v0.24.0 // indirect
 	go.opentelemetry.io/otel v1.11.1 // indirect
 	go.opentelemetry.io/otel/metric v0.33.0 // indirect
@@ -36,7 +34,8 @@ require (
 	google.golang.org/genproto v0.0.0-20211208223120-3a66f561d7aa // indirect
 	google.golang.org/grpc v1.51.0 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace github.com/epsagon/lupa/model => ../../../model
+
+replace github.com/epsagon/lupa/lupa-otelcol/internal/modeltranslator => ../../internal/modeltranslator
