--- lupa-otelcol/exporter/opensearchexporter/internal/modeltranslator/README.md
@@ -1,9 +1,0 @@
-# modeltranslator
-
-The `modeltranslator` package translates OTLP trace data to the internal model used by the Elasticsearch / OpenSearch exporters.
-
-## Usage
-
-```go
-internalSpans := modeltranslator.TranslateOTLPToInternalSpans(td.(ptrace.Traces))
-```
