+++ lupa-otelcol/internal/modeltranslator/model_translator.go
