+++ lupa-otelcol/internal/modeltranslator/model_translator_test.go
